refactor: consolidate player utils and lib into features/player
